requirements.txt
# Web framework
fastapi
uvicorn

# File upload handling
python-multipart

# OCR
easyocr
torch
torchvision

# PDF processing
PyMuPDF
Pillow

# Text processing
numpy

# HTTP client (for Ollama / APIs)
requests
